Reviewed the Coupons Plugin and Needs to Change.
This plugin should not exist like this at all.  The entire purpose of this plugin as is, is to redeem and claim credits.  If that is the case it should be nothing but an extra function or two in the credits plugin, with a few new fields needed in the credits table.

This entire plugin should be deleted and moved to something like, /credits/credits/redeem as the url, and /credits/credits/add to handle the adding of credits for a particular user to come and claim.

There will be a coupons plugin, but it needs to be part of the Orders plugin, and the model name will be something like OrderCoupon.  The relationship would be OrderTransaction hasMany or hasOne OrderCoupon, with the conditions of the coupon letting us know if one or more could be redeemed.  I'm going to begin work on the OrderCoupon addition to the Orders plugin now.